+++ index.js
@@ -1,44 +1,28 @@
 'use strict';
 
-const fromEntries = require('object.fromentries');
-const entries = require('object.entries');
-
 const configAll = require('./configs/all');
 const configRecommended = require('./configs/recommended');
 const configRuntime = require('./configs/jsx-runtime');
 
-const allRules = configAll.rules;
+const allRules = require('./lib/rules');
 
-function filterRules(rules, predicate) {
-  return fromEntries(entries(rules).filter((entry) => predicate(entry[1])));
-}
-
-function configureAsError(rules) {
-  return fromEntries(Object.keys(rules).map((key) => [`react/${key}`, 2]));
-}
-
-const activeRules = filterRules(allRules, (rule) => !rule.meta.deprecated);
-const activeRulesConfig = configureAsError(activeRules);
-
-const deprecatedRules = filterRules(allRules, (rule) => rule.meta.deprecated);
-
-const eslintrcPlugins = [
+// for legacy config system
+const plugins = [
   'react',
 ];
 
 module.exports = {
-  deprecatedRules,
+  deprecatedRules: configAll.plugins.react.deprecatedRules,
   rules: allRules,
   configs: {
     recommended: Object.assign({}, configRecommended, {
-      plugins: eslintrcPlugins,
+      plugins,
     }),
     all: Object.assign({}, configAll, {
-      plugins: eslintrcPlugins,
-      rules: activeRulesConfig,
+      plugins,
     }),
     'jsx-runtime': Object.assign({}, configRuntime, {
-      plugins: eslintrcPlugins,
+      plugins,
     }),
   },
 };